requirements.txt
streamlit
supabase-py
pandas
numpy
python-dotenv
